+++ src/domain/model/book/Author.java
@@ -1,4 +1,4 @@
-package domain.model;
+package domain.model.book;
 
 public record Author(String name) {
 }
